+++ config/karma.conf.js
@@ -96,7 +96,8 @@ module.exports = function (config) {
         colors: true,
         logLevel: config.LOG_INFO,
         autoWatch: false,
-      browsers: [ 'PhantomJS',
+      browsers: [
+        'PhantomJS',
         // 'Chrome'
         ],
         singleRun: true