+++ demo/app/components/tracks-layer/tracks-layer.component.ts
@@ -8,7 +8,8 @@ import {
   SimpleChanges,
   ViewChild
 } from '@angular/core';
-import { ConnectableObservable, Observable } from 'rxjs';
+import { Observable } from 'rxjs/Observable';
+import { ConnectableObservable } from 'rxjs/observable/ConnectableObservable';
 import { AcNotification } from '../../../../src/models/ac-notification';
 import { ActionType } from '../../../../src/models/action-type.enum';
 import { AcLayerComponent } from '../../../../src/components/ac-layer/ac-layer.component';
@@ -21,6 +22,7 @@ import { TracksDialogComponent } from './track-dialog/track-dialog.component';
 import { WebSocketSupplier } from '../../../utils/services/webSocketSupplier/webSocketSupplier';
 import { RealTracksDataProvider } from '../../../utils/services/dataProvider/real-tracks-data-provider';
 import { AppSettingsService } from '../../services/app-settings-service/app-settings-service';
+
 @Component({
   selector: 'tracks-layer',
   templateUrl: './tracks-layer.component.html',
@@ -195,18 +197,15 @@ export class TracksLayerComponent implements OnInit, OnChanges {
   }
 
   getPolylineColor() {
-    return new Cesium.Material({
-      fabric: {
-        type: 'Color',
-        uniforms: {
-          color: new Cesium.Color(0.3, 1.0, 0.3, 1.0)
-        }
-      }
-    });
+    return new Cesium.Color(0.3, 1.0, 0.3, 1.0);
   }
 
   showVelocityVectors(): boolean {
     return this.appSettingsService.showVelocityVectors;
+  }
+
+  showEllipses(): boolean {
+    return this.appSettingsService.showEllipses;
   }
 
   convertToCesiumObj(entity): any {