+++ src/components/ac-arc-desc/ac-arc-desc.component.ts
@@ -7,22 +7,24 @@ import { BasicStaticPrimitiveDesc } from '../../services/basic-primitive-desc/ba
 
 /**
  *  This is an implementation of an arc.
- *  The ac-arc-desc element must be a child of ac-layer element.
+ *  The element must be a child of ac-layer element.
  *  An arc is not cesium natively implemented and therefore it's API doesn't appear anywhere
+ *
  *  __Usage :__
  *  ```
- *    &lt;ac-arc-desc geometryProps="{
-            center: arc.center,
-            angle: arc.angle,
-            delta: arc.delta,
-            radius: arc.radius
-        }"
-        instanceProps="{
-            attributes: arc.attributes
-        }"
-        primitiveProps="{
-            appearance: arc.appearance"&gt;
- *    &lt;/ac-arc-desc&gt;
+ *    <ac-arc-desc geometryProps="{
+ *          center: arc.center,
+ *          angle: arc.angle,
+ *          delta: arc.delta,
+ *          radius: arc.radius
+ *       }"
+ *       instanceProps="{
+ *          attributes: arc.attributes
+ *       }"
+ *       primitiveProps="{
+ *          appearance: arc.appearance
+ *       }">
+ *    </ac-arc-desc>
  *    ```
  */