+++ src/components/ac-layer/ac-layer.component.ts
@@ -1,5 +1,14 @@
+// tslint:disable
 import { BillboardDrawerService } from '../../services/drawers/billboard-drawer/billboard-drawer.service';
-import { AfterContentInit, Component, Input, OnChanges, OnDestroy, OnInit, SimpleChanges } from '@angular/core';
+import {
+  AfterContentInit,
+  Component,
+  Input,
+  OnChanges,
+  OnDestroy,
+  OnInit,
+  SimpleChanges
+} from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { LayerService } from '../../services/layer-service/layer-service.service';
@@ -7,54 +16,74 @@ import { AcNotification } from '../../models/ac-notification';
 import { ActionType } from '../../models/action-type.enum';
 import { ComputationCache } from '../../services/computation-cache/computation-cache.service';
 import { LabelDrawerService } from '../../services/drawers/label-drawer/label-drawer.service';
-import { SimpleDrawerService } from '../../services/drawers/simple-drawer/simple-drawer.service';
-import { StaticCircleDrawerService } from '../../services/drawers/static-circle-drawer/static-circle-drawer.service';
 import { EllipseDrawerService } from '../../services/drawers/ellipse-drawer/ellipse-drawer.service';
-import { DynamicEllipseDrawerService } from '../../services/drawers/ellipse-drawer/dynamic-ellipse-drawer.service';
-import { DynamicPolylineDrawerService } from '../../services/drawers/dynamic-polyline-drawer/dynamic-polyline-drawer.service';
-import { StaticPolylineDrawerService } from '../../services/drawers/static-polyline-drawer/static-polyline-drawer.service';
-import { PolygonDrawerService } from '../../services/drawers/polygon-drawer/polygon-drawer.service';
+import { PolylineDrawerService } from '../../services/drawers/polyline-drawer/polyline-drawer.service';
 import { ArcDrawerService } from '../../services/drawers/arc-drawer/arc-drawer.service';
 import { PointDrawerService } from '../../services/drawers/point-drawer/point-drawer.service';
 import { AcEntity } from '../../models/ac-entity';
-
+import { BasicDrawerService } from '../../services/drawers/basic-drawer/basic-drawer.service';
+import { PolygonDrawerService } from '../../services/drawers/polygon-drawer/polygon-drawer.service';
+import { LayerOptions } from '../../models/layer-options';
+import { DynamicEllipseDrawerService } from '../../services/drawers/static-dynamic/ellipse-drawer/dynamic-ellipse-drawer.service';
+import { DynamicPolylineDrawerService } from '../../services/drawers/static-dynamic/dynamic-polyline-drawer/dynamic-polyline-drawer.service';
+import { StaticCircleDrawerService } from '../../services/drawers/static-dynamic/static-circle-drawer/static-circle-drawer.service';
+import { StaticPolylineDrawerService } from '../../services/drawers/static-dynamic/static-polyline-drawer/static-polyline-drawer.service';
+import { StaticPolygonDrawerService } from '../../services/drawers/static-dynamic/static-polygon-drawer/polygon-drawer.service';
+import { StaticEllipseDrawerService } from '../../services/drawers/static-dynamic/ellipse-drawer/ellipse-drawer.service';
+import { ModelDrawerService } from '../../services/drawers/model-drawer/model-drawer.service';
+// tslint:enable
 /**
  *  This is a ac-layer implementation.
  *  The ac-layer element must be a child of ac-map element.
- *  __param:__ {string} acfor - get the track observable and entityName (see the example)
- *  __param:__ {boolean} show - show/hide layer's entities
- *  __param:__ {any} context - get the context layer that will use the componnet (most of the time equal to "this")
+ *  __param:__ {string} acfor - get the track observable and entityName (see the example).
+ *  __param:__ {boolean} setShow - setShow/hide layer's entities.
+ *  __param:__ {any} context - get the context layer that will use the componnet (most of the time equal to "this").
+ *  __param:__ {LayerOptions} options - sets the layer options for each drawer.
  *
  *  __Usage :__
  *  ```
- *  &lt;ac-map&gt;
- *      &lt;ac-layer acFor="let track of tracks$" [show]="show" [context]="this"&gt;
- *          &lt;ac-billboard-desc props="{
- *               image: track.image,
- *               position: track.position,
- *               scale: track.scale,
- *               color: track.color,
- *               name: track.name
- *          }"&gt;
- *      &lt;/ac-billboard-desc>
- *          &lt;ac-label-desc props="{
- *               position: track.position,
- *               pixelOffset : [-15,20] | pixelOffset,
- *               text: track.name,
- *               font: '15px sans-serif'
- *          }"&gt;
- *          &lt;/ac-label-desc&gt;
- *      &lt;/ac-layer&gt;
- *  &lt;/ac-map&gt;
+ *  <ac-map>
+ *    <ac-layer acFor="let track of tracks$" [show]="show" [context]="this" [options]="options">
+ *      <ac-billboard-desc props="{
+ *        image: track.image,
+ *        position: track.position,
+ *        scale: track.scale,
+ *        color: track.color,
+ *        name: track.name
+ *      }">
+ *      </ac-billboard-desc>
+ *        <ac-label-desc props="{
+ *          position: track.position,
+ *          pixelOffset : [-15,20] | pixelOffset,
+ *          text: track.name,
+ *          font: '15px sans-serif'
+ *        }">
+ *      </ac-label-desc>
+ *    </ac-layer>
+ *  </ac-map>
  *  ```
  */
 @Component({
-  selector: 'ac-layer',
-  template: '',
-  providers: [
-    LayerService, ComputationCache, BillboardDrawerService, LabelDrawerService, EllipseDrawerService,
-    DynamicEllipseDrawerService, DynamicPolylineDrawerService, StaticCircleDrawerService,
-    StaticPolylineDrawerService, PolygonDrawerService, ArcDrawerService, PointDrawerService
+  selector : 'ac-layer',
+  template : '',
+  providers : [
+    LayerService,
+    ComputationCache,
+    BillboardDrawerService,
+    LabelDrawerService,
+    EllipseDrawerService,
+    PolylineDrawerService,
+    ArcDrawerService,
+    PointDrawerService,
+    PolygonDrawerService,
+    ModelDrawerService,
+    
+    DynamicEllipseDrawerService,
+    DynamicPolylineDrawerService,
+    StaticCircleDrawerService,
+    StaticPolylineDrawerService,
+    StaticPolygonDrawerService,
+    StaticEllipseDrawerService,
   ]
 })
 export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, OnDestroy {
@@ -66,52 +95,64 @@ export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, On
   context: any;
   @Input()
   store = false;
-
+  @Input()
+  options: LayerOptions;
+  
   private readonly acForRgx = /^let\s+.+\s+of\s+.+$/;
   private entityName: string;
   private stopObservable = new Subject();
   private observable: Observable<AcNotification>;
-  private _drawerList: SimpleDrawerService[] = [];
+  private _drawerList: Map<string, BasicDrawerService>;
   private _updateStream: Subject<AcNotification> = new Subject<AcNotification>();
   private entitiesStore = new Map<string, any>();
-
+  
   constructor(private  layerService: LayerService,
               private _computationCache: ComputationCache,
               billboardDrawerService: BillboardDrawerService,
               labelDrawerService: LabelDrawerService,
               ellipseDrawerService: EllipseDrawerService,
+              polylineDrawerService: PolylineDrawerService,
+              polygonDrawerService: PolygonDrawerService,
+              arcDrawerService: ArcDrawerService,
+              pointDrawerService: PointDrawerService,
+              modelDrawerService: ModelDrawerService,
+              
               dynamicEllipseDrawerService: DynamicEllipseDrawerService,
               dynamicPolylineDrawerService: DynamicPolylineDrawerService,
               staticCircleDrawerService: StaticCircleDrawerService,
               staticPolylineDrawerService: StaticPolylineDrawerService,
-              polygonDrawerService: PolygonDrawerService,
-              arcDrawerService: ArcDrawerService,
-              pointDrawerService: PointDrawerService) {
-    this._drawerList = Array.of(
-      billboardDrawerService,
-      labelDrawerService,
-      ellipseDrawerService,
-      dynamicEllipseDrawerService,
-      dynamicPolylineDrawerService,
-      staticCircleDrawerService,
-      staticPolylineDrawerService,
-      polygonDrawerService,
-      arcDrawerService,
-      pointDrawerService
-    );
+              staticPolygonDrawerService: StaticPolygonDrawerService,
+              staticEllipseDrawerService: StaticEllipseDrawerService) {
+    this._drawerList = new Map([
+      ['billboard', billboardDrawerService],
+      ['label', labelDrawerService],
+      ['ellipse', ellipseDrawerService],
+      ['polyline', polylineDrawerService],
+      ['polygon', polygonDrawerService as BasicDrawerService],
+      ['arc', arcDrawerService],
+      ['point', pointDrawerService],
+      ['model', modelDrawerService],
+      
+      ['dynamicEllipse', dynamicEllipseDrawerService],
+      ['dynamicPolyline', dynamicPolylineDrawerService],
+      ['staticCircle', staticCircleDrawerService],
+      ['staticPolyline', staticPolylineDrawerService],
+      ['staticPolygon', staticPolygonDrawerService],
+      ['staticEllipse', staticEllipseDrawerService],
+    ]);
   }
-
+  
   init() {
     this.initValidParams();
-
+    
     Observable.merge(this._updateStream, this.observable).takeUntil(this.stopObservable).subscribe((notification) => {
       this._computationCache.clear();
-
+      
       let contextEntity = notification.entity;
       if (this.store) {
         contextEntity = this.updateStore(notification);
       }
-
+      
       this.context[this.entityName] = contextEntity;
       this.layerService.getDescriptions().forEach((descriptionComponent) => {
         switch (notification.actionType) {
@@ -127,7 +168,7 @@ export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, On
       });
     });
   }
-
+  
   private updateStore(notification: AcNotification): any {
     if (notification.actionType === ActionType.DELETE) {
       this.entitiesStore.delete(notification.id);
@@ -145,12 +186,12 @@ export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, On
       }
     }
   }
-
+  
   private initValidParams() {
     if (!this.context) {
       throw new Error('ac-layer: must initialize [context] ');
     }
-
+    
     if (!this.acForRgx.test(this.acFor)) {
       throw new Error('ac-layer: must initialize [acFor] with a valid syntax \' [acFor]=\"let item of observer$\" \' '
         + 'instead received: ' + this.acFor);
@@ -162,33 +203,38 @@ export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, On
       throw  new Error('ac-layer: must initailize [acFor] with rx observable, instead received: ' + this.observable);
     }
   }
-
+  
   ngAfterContentInit(): void {
     this.init();
   }
-
+  
   ngOnInit(): void {
+    this._drawerList.forEach((drawer, drawerName) => {
+      const initOptions = this.options ? this.options[drawerName] : undefined;
+      drawer.init(initOptions);
+      drawer.setShow(this.show);
+    });
   }
-
+  
   ngOnChanges(changes: SimpleChanges): void {
-    if (changes['show']) {
+    if (changes.show && !changes.show.firstChange) {
       const showValue = changes['show'].currentValue;
       this._drawerList.forEach((drawer) => drawer.setShow(showValue));
     }
   }
-
+  
   ngOnDestroy(): void {
     this.stopObservable.next(true);
     this.removeAll();
   }
-
+  
   /**
    * Returns the store.
    */
   getStore(): Map<string, any> {
     return this.entitiesStore;
   };
-
+  
   /**
    * Remove all the entities from the layer.
    */
@@ -196,16 +242,16 @@ export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, On
     this.layerService.getDescriptions().forEach((description) => description.removeAll());
     this.entitiesStore.clear();
   }
-
+  
   /**
    * remove entity from the layer
    * @param {number} entityId
    */
   remove(entityId: string) {
-    this._updateStream.next({ id: entityId, actionType: ActionType.DELETE });
+    this._updateStream.next({id : entityId, actionType : ActionType.DELETE});
     this.entitiesStore.delete(entityId);
   }
-
+  
   /**
    * add/update entity to/from the layer
    * @param {AcNotification} notification
@@ -213,18 +259,18 @@ export class AcLayerComponent implements OnInit, OnChanges, AfterContentInit, On
   updateNotification(notification: AcNotification): void {
     this._updateStream.next(notification);
   }
-
+  
   /**
    * add/update entity to/from the layer
    * @param {AcEntity} entity
    * @param {number} id
    */
   update(entity: AcEntity, id: string): void {
-    this._updateStream.next({ entity, id, actionType: ActionType.ADD_UPDATE });
+    this._updateStream.next({entity, id, actionType : ActionType.ADD_UPDATE});
   }
-
+  
   refreshAll(collection: AcNotification[]): void {
-    // TODO make entity interface: collection of type entity not notification
-    Observable.from(collection).subscribe((entity) => this._updateStream.next(entity));
-  }
+		// TODO make entity interface: collection of type entity not notification
+		Observable.from(collection).subscribe((entity) => this._updateStream.next(entity));
+	}
 }