+++ src/models/entities-drawer-options.ts
@@ -1,0 +1,5 @@
+export interface EntitiesDrawerOptions {
+  collectionSuspensionTime: number;
+  collectionsNumber: number;
+  collectionMaxSize: number;
+}