+++ src/services/drawers/arc-drawer/arc-drawer.service.spec.ts
@@ -46,6 +46,10 @@ describe('ArcDrawerService', () => {
 			providers: [ArcDrawerService, providerFromMock(CesiumService, cesiumService)],
 		});
 	});
+  
+  beforeEach(inject([ArcDrawerService], (service: ArcDrawerService) => {
+    service.init();
+  }));
 
 	it('should create and return a new  arc primitive.', inject([ArcDrawerService], (service: ArcDrawerService) => {
 		const arc = service.add(arcProps.geometryProps, arcProps.instanceProps, arcProps.primitiveProps);